python_django/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-environ==0.11.2
psycopg2==2.9.10
filetype==1.2.0
django-filter==24.3
